fixed accidental nested test
